requirements.txt
stripe==2.56.0
requests==2.26.0
python-dotenv==0.19.1
python-telegram-bot==13.7
